Cmh generalize adaptive tau (#44)
Addresses #43 

## TL;DR; 
-- Currently a work in progress.  
-- Outstanding issue: Although the rate function argument can be
generalized and appears to be working, it does not work well in the
context of adaptivetau::ssa.exact() and troubleshooting is underway.
-- How to interact with this PR: Two example scripts are now in the new
compModels/R/scripts folder to demonstrate what is going on. One
generalization that works for SIR model using sir_rates as rateFunc in
the script named run_gen_stochastic_tau_sir.r. For using
generalized_rates as the rateFunc, see the script
generalizing_rates_tau.r. Some troubleshooting I have attempted in this
latter script includes sending in named lists and trying to follow the
object requirements in the ssa.exact() documentation but this has not
resolved it.

## Background and context
I first developed a runner function run_sir_stochastic_tau that could
run a stochastic SIR model using the ssa.exact() function in the
adaptivetau package. You can find this function
[here](https://github.com/cdcent/cfa-gen-comp-model-tools/blob/main/compModels/R/run_sir_stochastic_tau.R).
This function makes use of the following helpers:

[validate_sir_stoch_input()](https://github.com/cdcent/cfa-gen-comp-model-tools/blob/main/compModels/R/validate_sir_stoch_input.R)

[sir_rates()](https://github.com/cdcent/cfa-gen-comp-model-tools/blob/main/compModels/R/sir_rates.R)

In this PR #44 I try to generalize run_sir_stochastic_tau to be able to
handle a user defined number and name of compartments. The rational for
using the adaptivetau package was to align with Measles response work.
 
run_sir_stochastic_tau function was one of three (deterministic,
stochastic using GillespieSSA::ssa(), stochastic using adaptivetau
package) I first made many months ago when using a concrete
compartmental model example as part of development of the compModels
package. Since then, I have worked to generalize the input on all three
functions and the package currently has the [deterministic
](https://github.com/cdcent/cfa-gen-comp-model-tools/blob/main/compModels/R/run_gen_deterministic.R)
and [stochastic model using Gillespie SSA
package](https://github.com/cdcent/cfa-gen-comp-model-tools/blob/main/compModels/R/run_gen_stochastic.r)
working. By working I mean that it can take a user defined number and
name of compartments and adjust and run without error and produce plots.
These functions also have input and output validation functions and some
tests of performance. Scientific testing is still minimal at this stage
(eg final size and initial growth tests/functions available).
Additionally, the process of generalizing lead to the development of
several additional helper functions to handle the expansion of the
transmission matrices and application of multiplier values/matrices to
modify transitions during interventions.
 
## Where I think the outstanding issue lies:
In this PR I attempt to generalize the run_sir_stochastic_tau.r function
to be able take a user defined number and name of compartments and am
running into trouble in some, but not all, use cases. The issue appears
to be in the rateFunc argument of adaptivetau::ssa.exact() and how that
package uses what is supplied to it. For the SIR model I developed
first, I fed it sir_rates and this worked well. For the first pass at
generalizing the function, I fed in the function generalized_rates but
asked it to handle a SIR case, and this also worked. Both of these can
be found in the script run_gen_stochastic_tau_sir.r. But when I tried to
do something other than a SIR, I run into several errors which are
detailed in the order I tried to resolve them in the script
generalizing_rates_tau.r. I have added many print statements into
generalized_rates() to try to catch where the issue lies.

Although generalized_rates() runs well by itself and produces the
expected output, when it is used in the rateFunc argument in
adaptivetau::ssa.exact() inside the wrapper function
run_gen_stochastic_tau(), something happens to the objects that I have
not figured out. For example, the print statements show that parameters
are in the in the state object instead of states and nothing is in the
params object.

The part of generalized_rates that might be interacting poorly may be
one of the following lines using list2env() and eval():
1.   eval_env <- list2env(eval_env_list)
2.   rates[idx] <- eval(parse(text = rate_expr), envir = eval_env)

---------

Co-authored-by: Beau B. Bruce <lue7@cdc.gov>