First Pass draft of Chicago Measles Response 2024 modeling work (#96)
This script serves as a first pass working draft of Chicago Measles
Response 2024 modeling work, and a list of functionality that remains to
be developed to completely recreate the modeling work, along with other
future useful directions.

Code was prepared during paired coding between @cherz4 and
@bradfordptaylor and each contributed to the final file here.

Additionally, a few other files were changed as the merge-two-repos
branch is behind updates on main for a few files.