Issues #64 Subtask 3 Remove validation scripts and check scripts that will not be used/cannot be modified (#109)
Addresses Issue #64 Subtask 3 Remove validation scripts and check
scripts that will not be used/cannot be modified

Specifically it removes:

1. All 4 remaining validation functions, as they were function specific
to functions that were part of the old compModels setup. Further
dicussion is needed if we want to make validation functions (that
combine a subset of the check functions specific to the function being
validated) to accompany each of the compModels functions or if this
should be in the testing suite.

2. Several check_ functions, specifically:
- check_gillespie_args() – wrapper takes compiled model which provides
transitions matrix
- check_adaptivetau_args() – wrapper takes compiled model which has the
change matrix and propensity fns
- check_modifier1() - modifiers for interventions are not currently
working the same in new compModels
- check_modifier2() - modifiers for interventions are not currently
working the same in new compModels
- check_subgroups_list() - no longer needed as we don’t have same
subgroups object in compiled model
- check_trans_mtx() - checked that the row and column names were
compartment names. We don’t have the same transmission matrix object
anymore.


The following will remain as they only need little or no modification.
Needs modification: 
- check_init_vals_compnames() – this checked the length was the same for
names and values – this check may already exist in test suite. This must
be modified
- check_intervention_times() - checks there are single numeric start and
end values (or that they are NULL). Unclear if NULL option should be
removed

No modification needed:  
- check_init_vals() - can be applied to any wrapper taking initial vals 
- check_nsims() - checks # time steps are single positive numeric value,
should work in any function taking # of sims
- check_parms_vec() - checks that parameter values are numeric and
positive. Should work in any function taking parameters if these are
desired parameter definitions
- check_time_step() - checks timesteps are a single positive numeric
value, should work in any function taking a single value for time steps
- check_times() - checks that that a time vector has a least one numeric
element and is strictly increasing. Could be applied to any time vector
taken in by a function that is supposed to be increasing