Issue #90 - fixes rigid init_vals input fixed on simulation wrappers (#167)
Closes Issue #90 - fixes rigid init_vals input fixed on simulation
wrappers

- Removed use of define_states() and output_initialstate() inside the
simulation wrappers, per suggestion on Issue #90 . We expect user to
input properly formatted initial conditions to the wrapper, not have it
wrangled on the inside. This had already been done for wrap_adaptivetau
so this brings the other two fns up to date with this change.
- Minor Roxygen tweaks on wrap_gillespie and wrap_adaptivetau first
examples to assign output to an object (or it doesn't show up)