Cleaning up no longer used functions & remaining old compModels function Roxygen (Issue #132) (#168)
Resolves Issue #132 
Removed three fns that are no longer used in the code base. They are
working and could be useful but are not used and are from old
compModels.
  deleted:    compModels/R/calculate_cuminf.R
  deleted:    compModels/R/is_intervention_period.R
deleted: compModels/R/sir_final_size.R - we have a different final size
check in our test suite test-sir-homogeneous-final-size.R

Kept but modified: compModels/R/generalized_rates.R to remove the
comment banner about this function possibly being removed/modified in
the SIRmodelbuilder merge. This did not occur and this function is still
referenced by our simulation functions but it is used as the default
when a user does not specific another function to use.